requirements.txt
fastapi==0.115.4
joblib==1.4.2
keras==3.4.1
numpy
openai==0.28.0
pandas
python-dotenv==1.0.1
scikit-learn==1.5.1
sentence-transformers==3.0.1
sentencepiece==0.2.0
tensorflow==2.17.0
tf_keras==2.17.0
transformers==4.43.3
uvicorn==0.32.0
boto3
b2sdk
